use long queue on stoomboot

+++ core/perform_simlations_grid.py
@@ -40,7 +40,7 @@ N = {N}
 
 def make_grid(N_grid = 10, D_grid = 10):
     # Create a cluster object with a square grid of stations
-    
+
     transform = lambda x : D_grid * x - N_grid/2 * D_grid
     cluster = BaseCluster()
 
@@ -87,7 +87,7 @@ def perform_simulations():
             print "/data/hisparc/corsika/data/{seeds}/corsika.h5 does not exist".format(seeds=seed)
             continue
 
-        perform_job(seed, 'short')
+        perform_job(seed, 'long')
 
     query.finish()
